Use Meziantou analyzers
