+++ src/dotnet-decode-jwt/Infrastructure/NonThreadSafeConsole.cs
@@ -1,14 +1,5 @@
 namespace Gabo.DotNet.DecodeJwt.Infrastructure;
 
-public interface IConsole
-{
-    void WriteBoringLine(string value);
-    void WriteDullLine(string value);
-    void WriteInfoLine(string value);
-    void WriteFancyLine(string value);
-    void WriteErrorLine(string value);
-}
-
 internal class NonThreadSafeConsole : IConsole
 {
     internal const ConsoleColor BoringColor = ConsoleColor.Gray;
